d-basic: include c-basic
The `ldc` package is patched to use `ld.bfd` by default on Clear Linux
OS, and it would otherwise default to `ld.gold`. Both of these linkers
reside in `binutils`. Other gcc/glibc development files are needed as
well, so including c-basic is the easiest approach to more fully satisfy
runtime dependencies for this bundle.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>