Deprecate kernels earlier than ltsprev (lts2022)
We've already modified the bundles for older LTS kernels to include
ltsprev, along with the clr-boot-manager change to migrate boot kernels
from older LTS to ltsprev. Now it's time to remove those older kernels
and mark the bundles deprecated so we can remove them in a future format
bump.

Also make sure the dkms bundles simply include kernel-ltsprev-dkms, and
remove the supplemental initrds for the older LTS kernels.