Starting bundle cleanup
This change is removing bundles for which there is a pundle providing
the same name. In some cases content may be dropped from the bundle
and in other cases dependencies will be added to the package. If there
is content missing from the new pundle it probably should be added
back in some way. For now take the biggest offenders for include
breakage though as well as y-z pundle names.

Signed-off-by: William Douglas <william.douglas@intel.com>